pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "voxd"
version = "1.3.1"  # bump manually on releases
description = "Voice-typing helper powered by whisper.cpp"
authors = [{ name = "Jakov", email = "jakov.iv@proton.me" }]
requires-python = ">=3.9"
dependencies = [
  "sounddevice>=0.5",
  "pyqt6>=6.9",
  "platformdirs>=4.2",
  "pyyaml>=6.0",
  "pyperclip>=1.9",
  "psutil>=5.9",
  "numpy>=1.26",
  "requests>=2.28.0",
  "tqdm>=4.66",
  "pyqtgraph>=0.13"
]


[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["llama.cpp","whisper.cpp","venv",".venv","build","dist"]

[project.scripts]
voxd = "voxd.__main__:main"
voxd-model = "voxd.models:_cli"

[tool.hatch.build]
package-dir = "src"
include = ["setup.sh","src/voxd/assets/*.png","src/voxd/defaults/*.yaml","launcher_setup.sh"]

[tool.hatch.build.targets.wheel]
packages = ["src/voxd"]

[tool.hatch.build.targets.editable]
packages = ["src/voxd"]
